Delegate node moving to Spline
This lets invalid moves be detected and rejected, but it
doesn't actually work because once the mouse moves past the
invalid range then the point just snaps over the obstacle.
Can't just disable the drag and drop since then you might
not be able to get the node back out again.